+++ Basil/include/Renderer/Camera.h
@@ -11,9 +11,11 @@ namespace Basil
 	{
 		public:
 			// TODO: move definition to own cpp file
+			Camera() = default;
 			Camera(const glm::mat4& projection) : projection(projection) {}
+			virtual ~Camera() = default;
 			const glm::mat4& getProjection() const { return projection; }
-		private:
-			glm::mat4 projection;
+		protected:
+			glm::mat4 projection = glm::mat4(1.0f);
 	};
 }