+++ Basil/include/Scene/Scene.h
@@ -19,8 +19,11 @@ namespace Basil
 			Entity createEntity(const std::string& name = std::string());
 			entt::registry& reg();
 			void onUpdate(Timestep timeStep);
+			void onViewportResize(uint32_t width, uint32_t height);
 		private:
 			entt::registry registry;
+			uint32_t viewportWidth;
+			uint32_t viewportHeight;
 			friend class Entity;
 	};
 }